Introduce a mechanism for robust corpus tests
Including all of these JDK sources here (with overwritten copyright headers!) is critical. This changeset removes all JDK sources and instead replaces it with a more robust (and legal) corpus test system.

The corpus consists of open-source libraries (currently only `monticore-runtime` but this will change soon). The source artifacts of these are downloaded using Gradle's dependency mechanisms. To customize the corpus, simply add/remove dependencies to the "corpus" configuration in the build script.